% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burritos_cost.R
\name{burritos_cost}
\alias{burritos_cost}
\title{Calculate Total Burrito Cost}
\usage{
burritos_cost(price, count)
}
\arguments{
\item{price}{Price per burrito in dollars}

\item{count}{Price per burrito in dollars}
}
\description{
Calculate Total Burrito Cost
}
